Texts updated; Summary.java - fixed bug in createImage() within run(); SummItem.java - decorations; join.pl, join_common.pl - added more exclusions; mutter/* - aphetics continued; release_gui.pl - state clean from $

+++ Astromaximum/mutter/evclass.cpp
@@ -30,7 +30,7 @@ int Event::getDayNumber()
   return (julianDay-startJD)/MINUTE_STEP;
 }
 
-int Event::packDate(double date)
+long Event::packDate(double date)
 {
   int y,m,d; double hour;
   swe_revjul(date, SE_GREG_CAL, &y, &m, &d, &hour);
